+++ sql/core/src/test/scala/org/apache/spark/sql/streaming/StreamSuite.scala
@@ -20,12 +20,10 @@ package org.apache.spark.sql.streaming
 import scala.reflect.ClassTag
 import scala.util.control.ControlThrowable
 
-import org.apache.spark.SparkException
 import org.apache.spark.sql._
 import org.apache.spark.sql.execution.streaming._
 import org.apache.spark.sql.sources.StreamSourceProvider
 import org.apache.spark.sql.types.{IntegerType, StructField, StructType}
-import org.apache.spark.util.ManualClock
 
 class StreamSuite extends StreamTest {
 
@@ -300,48 +298,6 @@ class StreamSuite extends StreamTest {
       assert("LocalTableScan".r.findAllMatchIn(explainWithExtended).size === 1)
     } finally {
       q.stop()
-    }
-  }
-
-  test("StreamingQuery should be Serializable but cannot be used in executors") {
-    def startQuery(ds: Dataset[Int], queryName: String): StreamingQuery = {
-      ds.writeStream
-        .queryName(queryName)
-        .format("memory")
-        .start()
-    }
-
-    val input = MemoryStream[Int]
-    val q1 = startQuery(input.toDS, "stream_serializable_test_1")
-    val q2 = startQuery(input.toDS.map { i =>
-      // Emulate that `StreamingQuery` get captured with normal usage unintentionally.
-      // It should not fail the query.
-      q1
-      i
-    }, "stream_serializable_test_2")
-    val q3 = startQuery(input.toDS.map { i =>
-      // Emulate that `StreamingQuery` is used in executors. We should fail the query with a clear
-      // error message.
-      q1.explain()
-      i
-    }, "stream_serializable_test_3")
-    try {
-      input.addData(1)
-
-      // q2 should not fail since it doesn't use `q1` in the closure
-      q2.processAllAvailable()
-
-      // The user calls `StreamingQuery` in the closure and it should fail
-      val e = intercept[StreamingQueryException] {
-        q3.processAllAvailable()
-      }
-      assert(e.getCause.isInstanceOf[SparkException])
-      assert(e.getCause.getCause.isInstanceOf[IllegalStateException])
-      assert(e.getMessage.contains("StreamingQuery cannot be used in executors"))
-    } finally {
-      q1.stop()
-      q2.stop()
-      q3.stop()
     }
   }
 }
