+++ sql/core/src/test/scala/org/apache/spark/sql/streaming/StreamingQuerySuite.scala
@@ -17,7 +17,6 @@
 
 package org.apache.spark.sql.streaming
 
-import java.util.{Collections, UUID}
 import java.util.concurrent.CountDownLatch
 
 import org.apache.commons.lang3.RandomStringUtils
@@ -27,7 +26,7 @@ import org.scalatest.BeforeAndAfter
 import org.scalatest.concurrent.PatienceConfiguration.Timeout
 
 import org.apache.spark.internal.Logging
-import org.apache.spark.sql.DataFrame
+import org.apache.spark.sql.{DataFrame, Dataset}
 import org.apache.spark.sql.types.StructType
 import org.apache.spark.SparkException
 import org.apache.spark.sql.execution.streaming._
@@ -440,35 +439,46 @@ class StreamingQuerySuite extends StreamTest with BeforeAndAfter with Logging {
     }
   }
 
-  test("progress classes should be Serializable") {
-    // Fake the objects and verify that they can be serialized/deserialized.
-    val stateOperatorProgress = new StateOperatorProgress(numRowsTotal = 10, numRowsUpdated = 10)
-    val sourceProgress = new SourceProgress(
-      description = "test",
-      startOffset = "test-offset",
-      endOffset = "test-offset",
-      numInputRows = 1,
-      inputRowsPerSecond = 0.1,
-      processedRowsPerSecond = 0.1
-    )
-    val sinkProgress = new SinkProgress(description = "test")
-    val streamingQueryProgress = new StreamingQueryProgress(
-      id = UUID.randomUUID(),
-      runId = UUID.randomUUID(),
-      name = "test-query",
-      timestamp = "2016-12-05T20:54:20.827Z",
-      batchId = 100,
-      durationMs = Collections.singletonMap("test", 10L),
-      eventTime = Collections.singletonMap("watermark", "10"),
-      stateOperators = Array(stateOperatorProgress),
-      sources = Array(sourceProgress),
-      sink = sinkProgress
-    )
-    val array = spark.sparkContext.parallelize(Seq(streamingQueryProgress)).collect()
-    assert(array.length === 1)
-    // Make sure we did serialize and deserialize the object
-    assert(array(0) ne streamingQueryProgress)
-    assert(array(0).json === streamingQueryProgress.json)
+  test("StreamingQuery should be Serializable but cannot be used in executors") {
+    def startQuery(ds: Dataset[Int], queryName: String): StreamingQuery = {
+      ds.writeStream
+        .queryName(queryName)
+        .format("memory")
+        .start()
+    }
+
+    val input = MemoryStream[Int]
+    val q1 = startQuery(input.toDS, "stream_serializable_test_1")
+    val q2 = startQuery(input.toDS.map { i =>
+      // Emulate that `StreamingQuery` get captured with normal usage unintentionally.
+      // It should not fail the query.
+      q1
+      i
+    }, "stream_serializable_test_2")
+    val q3 = startQuery(input.toDS.map { i =>
+      // Emulate that `StreamingQuery` is used in executors. We should fail the query with a clear
+      // error message.
+      q1.explain()
+      i
+    }, "stream_serializable_test_3")
+    try {
+      input.addData(1)
+
+      // q2 should not fail since it doesn't use `q1` in the closure
+      q2.processAllAvailable()
+
+      // The user calls `StreamingQuery` in the closure and it should fail
+      val e = intercept[StreamingQueryException] {
+        q3.processAllAvailable()
+      }
+      assert(e.getCause.isInstanceOf[SparkException])
+      assert(e.getCause.getCause.isInstanceOf[IllegalStateException])
+      assert(e.getMessage.contains("StreamingQuery cannot be used in executors"))
+    } finally {
+      q1.stop()
+      q2.stop()
+      q3.stop()
+    }
   }
 
   /** Create a streaming DF that only execute one batch in which it returns the given static DF */
